cr changes

+++ Source/Relativity.Transfer.SDK.Samples.Repository/FullPathWorkflow/BearerTokenAuthentication.cs
@@ -15,7 +15,7 @@ using Relativity.Transfer.SDK.Samples.Core.Runner;
 
 namespace Relativity.Transfer.SDK.Samples.Repository.FullPathWorkflow;
 
-[Sample((int)SampleOrder.BearerTokenAuthentication, "A bearer token authentication",
+[Sample(SampleOrder.BearerTokenAuthentication, "A bearer token authentication",
 	"The sample illustrates how to implement a bearer token authentication in RelativityOne.",
 	typeof(BearerTokenAuthentication),
 	TransferType.UploadFile)]
